Week's developed Features (#57)
## Weekly progress update 04/08/2023

This is the Weekly progress update on the development of the web-based IP Reputation monitoring tool. Here's what we accomplished this week:

#### Front-End Updates:

We linked the lists of blocked IP addresses from CINS and blocklist to the front end. The system can now search for the Organisation’s IP addresses from the list and return found IPs while indicating from which site/Database they have been found.

### GitHub Actions:

We are setting up Github actions to streamline our development process and ensure smoother deployments, currently we have an action workflow where tests are run on making a pull request to the `main` and `development` branches and we planned on how to add the code quality tests and automated deployments, CI/CD pipeline.

### Testing and Code Quality:

Installed code coverage and Flake8 for testing and maintaining the code quality following the [PEP8 guidelines](https://pep8.org/). This will help us catch any potential issues early in the development cycle.

### Add User Feature:

We worked on implementing the "Add User" feature. Users can now be added to the system, which will be an essential part of access control and user management.

### Prototype in Figma:

We created a prototype for the User Interface showing the flow of the system in Figma. This will serve as a visual reference for our design and development decisions and make the realization of the frontend guided and faster.

### Feature-Based Branches:

We modified our branch naming to be feature-based and came up with a contributing guideline documentation for the developers. This will help us organize our work better and enable easier collaboration between team members.

### Back-End Progress:

We identified RENU's IPv4 and IPv6 subnets from the Spamhaus list. This is a significant step towards populating the list with relevant data for monitoring.

### Pentesting Team Feedback:

We are working closely with the Pentesting Team and have addressed all the suggestions they made from our previous development. We have cleaned up the .gitignore file and are making sure none of the secret information is under public version control.